requirements.txt
torch==2.6.0
pandas==2.2.3
scikit-learn==1.4.2
optuna==3.4.0
matplotlib==3.8.0
numpy==1.26.0
pyyaml==6.0.1